+++ inputmask-pages/src/assets/Changelog.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 
+- Optional parts not working as expected #2710
 - State of alternator not reset with clearIncomplete = true #2712
 - Change Event fired on Blur even if value has not changed #2691
 - fix ssr issues when navigator is not defined #2704